requirements.txt
Django>=2.0,<3.0
psycopg2-binary>=2.8
djangorestframework==3.11.0

